Switched from using Boost::multi_array to using a wrapper for a 1D array to 3D conversion.
Operator overloaded for both MyVector and array3d classes.
Operators that work now are: +,-,*,/ and are defined for interoperability between MyVector, array3d, and float. C++ should cast correctly (I think) so int can be used in place of float.

Tested the output for each of the overloads.

TODO: Get the tuple thing working to return the shape of the array to do assertions correctly.